% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{r_image}
\alias{r_image}
\title{Image}
\usage{
r_image(r, src)
}
\arguments{
\item{r}{A \code{raphael} object as returned by \code{\link{raphael}}}

\item{src}{the image source}
}
\description{
Add an image.
}
\examples{

raphael() \%>\%
 r_image(src = "https://dmitrybaranovskiy.github.io/raphael/bd.jpg")

}
