% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{r_pie}
\alias{r_pie}
\title{Pie chart}
\usage{
r_pie(r, data, x, y)
}
\arguments{
\item{r}{A \code{raphael} object as returned by \code{\link{raphael}}}

\item{data}{data.frame containing data to plot.}

\item{x}{variable column.}

\item{y}{variable column.}
}
\description{
Add a pie chart.
}
\examples{
library(dplyr)
library(tibble)

raphael() \%>\%
 r_pie(mtcars \%>\%
         rownames_to_column("names") \%>\%
         arrange(-hp) \%>\%
         head(5),
       x = names,
       y = hp)

}
