% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raphael.R
\name{raphael}
\alias{raphael}
\title{Initiate a raphael.}
\usage{
raphael(width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{width, height}{dimensions of chart.}

\item{elementId}{id of div containing chart.}
}
\description{
Initiate a raphael graph.
}
\examples{
library(dplyr)
library(tibble)

raphael() \%>\%
 r_pie(mtcars \%>\%
         rownames_to_column("names") \%>\%
         arrange(-hp) \%>\%
         head(5),
       x = names,
       y = hp)

}
