% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raphael.R
\name{raphael}
\alias{raphael}
\title{Initiate a raphael.}
\usage{
raphael(
  data = NULL,
  x = NULL,
  y = NULL,
  src = NULL,
  type = "pie",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{data.frame containing data to plot.}

\item{x}{variable column.}

\item{y}{variable column.}

\item{src}{image source.}

\item{type}{type of chart.}

\item{width, height}{dimensions of chart.}

\item{elementId}{id of div containing chart.}
}
\description{
Initiate a raphael graph.
}
\examples{
library(dplyr)
library(tibble)

mtcars \%>\% rownames_to_column("names") \%>\% raphael(x = names, y = carb, type = "pie")

}
